+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Improve compare configs for ProductSearch and PDP
+
 ## [1.97.0] - 2026-03-19
 
 ### Changed